% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalaConvert.featureAllocation.R
\name{scalaConvert.featureAllocation}
\alias{scalaConvert.featureAllocation}
\title{\strong{(Developers Only:)} Convert Between R and Scala Representations of Feature Allocations}
\usage{
scalaConvert.featureAllocation(x, names = NULL, withParameters = TRUE)
}
\arguments{
\item{x}{Either: i. a binary matrix or list of binary matrices encoding feature allocation(s).
ii. a Scala reference to a feature allocation or an array of Scala references to
feature allocations.}

\item{names}{A character vector giving the row names when converting from Scala to R.}

\item{withParameters}{A logical indicating whether model parameters should also be converted.}
}
\description{
\strong{This function is not intended for end users}, but is exported for the benefit of
developers whose wish to write other packages that depend on this package.
}
