% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USArrests.featureAllocations-data.R
\docType{data}
\name{USArrests.featureAllocations}
\alias{USArrests.featureAllocations}
\title{Feature Allocations of the USArrests Dataset}
\format{A list of 1000 randomly generated feature allocations for the 50 states in the USArrests dataset.}
\usage{
USArrests.featureAllocations
}
\description{
Randomly generated feature allocations of the states in the USArrests dataset.
}
\seealso{
\code{\link[datasets]{USArrests}}
}
\keyword{datasets}
